requirements.txt
fastapi==0.115.4
scrapy==2.11.2
selenium==4.26.1
webdriver_manager==4.0.2
beautifulsoup4==4.12.3
uvicorn==0.32.0
opencv-python==4.10.0.84
numpy==2.0.0
elasticsearch-dsl==8.15.3